gpu: nvgpu: fix compile error of new compile flags
Preparing to push hvrtos gpu server changes which requires bellow CFLAGS:
        -Werror -Wall -Wextra \
        -Wmissing-braces -Wpointer-arith -Wundef \
        -Wconversion -Wsign-conversion \
        -Wformat-security \
        -Wmissing-declarations -Wredundant-decls -Wimplicit-fallthrough

Jira GVSCI-11640

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvgpu/+/2653746
GVS: Gerrit_Virtual_Submit

+++ drivers/gpu/nvgpu/common/gr/fecs_trace.c
@@ -515,7 +515,7 @@ int nvgpu_gr_fecs_trace_poll(struct gk20a *g)
 
 	read = g->ops.gr.fecs_trace.get_read_index(g);
 
-	cnt = CIRC_CNT(write, read, GK20A_FECS_TRACE_NUM_RECORDS);
+	cnt = CIRC_CNT((u32)write, (u32)read, GK20A_FECS_TRACE_NUM_RECORDS);
 	if (!cnt)
 		goto done;
